feat(errorprone): enable IntLongMath
Manager.getPendingSize widens its first queue length, so the three
counts add up in long arithmetic. That one is a node-local metric.

The other three sites are frozen and annotated instead. The two
precompile energy formulas are what mainnet charges, down to the legacy
negative count for undersized calldata, and the create-account size
check in BandwidthProcessor decides TooBigTransactionException for
in-block transactions once ALLOW_CONSENSUS_LOGIC_OPTIMIZATION is active.
Each annotation records why the int expression cannot overflow today.

+++ framework/src/main/java/org/tron/core/db/Manager.java
@@ -2552,7 +2552,7 @@ public class Manager {
   }
 
   public long getPendingSize() {
-    long value = getPendingTransactions().size() + getRePushTransactions().size()
+    long value = (long) getPendingTransactions().size() + getRePushTransactions().size()
         + getPoppedTransactions().size();
     return value;
   }
